requirements.txt
--index-url https://download.pytorch.org/whl/cu128
torch==2.9.1
torchvision==0.24.1
--index-url https://pypi.org/simple
open_clip_torch==3.2.0
timm==1.0.24
numpy==2.4.1
pillow==12.1.0
tqdm==4.67.1
opencv-python==4.13.0.90
scikit-learn==1.8.0
matplotlib==3.10.8
setuptools==70.0.0
einops
pycocotools
psutil
decord
